update array length decode
